Support Instances of Allowed R5 Resources in R4/R4B IGs
Allow authors to create instances of R5 ActorDefinition, Requirements, SubscriptionTopic, and TestPlan in R4 IGs - and refer to them via references and canonicals (e.g., from an obligations extension).

This is supported by re-using the "supplemental FHIR packages" capability of FHIRDefinitions (previously used only by cross-version extensions) and implementing a new way to fish that prefers a version of FHIR different from the IG's declared fhirVersion. Now, when an R4 IG is loaded, R5 will automatically be loaded as a supplemental FHIR version since authors do not declare ahead of time if they will be using these special resources.

Other changes to support this: MasterFisher now determines its default version of FHIR when it is initialized. I've also added a FHIRVersionsUtility and refactored some existing code to leverage it.